+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Parse Markdown files instead of HTML files.
+
 ## 0.7 (2023-03-14)
 
 - Detect Cloze deletions with HTML inside, e.g.: {{c1::<sup>14</sup>C}}